pyproject.toml
[project]
name = "extraction-ai"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "hathimamir", email = "hathim711@gmail.com" }
]
dependencies = [
    "Pillow==10.4.0",
    "PyMuPDF==1.24.10",
    "accelerate==0.24.1",
    "albumentations==1.4.18",
    "datasets==2.14.6",
    "evaluate==0.4.1",
    "mistralai>=0.4.0",
    "numpy==2.1.2",
    "openpyxl==3.1.5",
    "pandas==2.2.3",
    "pdf2image==1.17.0",
    "pdfplumber==0.11.4",
    "pytest==8.3.3",
    "python-dotenv>=1.0.0",
    "pyyaml==6.0.2",
    "torch==2.4.1",
    "transformers==4.45.2",
    "pydantic==2.7.1",
    "reportlab==4.2.0",
    "faker==25.2.0",
    "jinja2==3.1.4",
    "weasyprint==62.0",
    "pypdf2==3.0.1",
]
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "sentencepiece>=0.2.0",
    "protobuf>=6.31.1",
    "typer>=0.12.3",
    "rich>=13.7.1",
    "ruff>=0.4.4",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/extraction_ai"]
